Stats: Devices module - moving files to new directory
