+++ client/my-sites/stats/features/modules/stats-devices/stats-module-devices.scss
@@ -1,5 +1,5 @@
 @import "@wordpress/base-styles/breakpoints";
-@import "../modernized-mixins";
+@import "../../../modernized-mixins";
 
 .stats-module-devices {
 	@include segmented-controls;
